+++ src/Phan/Analysis/PostOrderAnalysisVisitor.php
@@ -9,6 +9,7 @@ use ast\Node;
 use Closure;
 use Exception;
 use Phan\AST\AnalysisVisitor;
+use Phan\AST\ASTReverter;
 use Phan\AST\ASTSimplifier;
 use Phan\AST\ContextNode;
 use Phan\AST\PhanAnnotationAdder;
@@ -949,6 +950,13 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
                 Issue::NoopCast,
                 $node->lineno,
                 self::AST_CAST_FLAGS_LOOKUP[$node->flags] ?? 'unknown'
+            );
+        }
+        if ($node->flags === flags\TYPE_NULL) {
+            $this->emitIssue(
+                Issue::CompatibleUnsetCast,
+                $node->lineno,
+                ASTReverter::toShortString($node)
             );
         }
         return $this->context;
@@ -1942,32 +1950,9 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
                         $node->lineno,
                         (string)$class->getFQSEN(),
                         $class->getContext()->getFile(),
-                        (string)$class->getContext()->getLineNumberStart()
+                        (string)$class->getContext()->getLineNumberStart(),
+                        $class->getDeprecationReason()
                     );
-                }
-                foreach ($class->getInterfaceFQSENList() as $interface) {
-                    $clazz = $this->code_base->getClassByFQSEN($interface);
-                    if ($clazz->isDeprecated()) {
-                        $this->emitIssue(
-                            Issue::DeprecatedInterface,
-                            $node->lineno,
-                            (string)$clazz->getFQSEN(),
-                            $clazz->getContext()->getFile(),
-                            (string)$clazz->getContext()->getLineNumberStart()
-                        );
-                    }
-                }
-                foreach ($class->getTraitFQSENList() as $trait) {
-                    $clazz = $this->code_base->getClassByFQSEN($trait);
-                    if ($clazz->isDeprecated()) {
-                        $this->emitIssue(
-                            Issue::DeprecatedTrait,
-                            $node->lineno,
-                            (string)$clazz->getFQSEN(),
-                            $clazz->getContext()->getFile(),
-                            (string)$clazz->getContext()->getLineNumberStart()
-                        );
-                    }
                 }
             }